제목 : 네비게이터 추가 * feature : 새로운 기능
1. 네비게이터 추가
2. 홈 화면 홈 파일로 이동
3. 홈 화면에 시간표로 이동하는 버튼 추가
