+++ Front-end/LimeFront/TimeSchedule.js
@@ -1,0 +1,10 @@
+import React, {Component} from "react";
+import {Text} from "react-native";
+
+export default class TimeSchedule extends Component {
+    render() {
+        return (
+            <Text>시간표</Text>
+        )
+    }
+}
